+++ clang/lib/Parse/ParseExprCXX.cpp
@@ -2098,8 +2098,17 @@ Parser::ParseCXXCondition(StmtResult *InitStmt, SourceLocation Loc,
       return ParseCXXCondition(nullptr, Loc, CK, MissingOK);
     }
 
-    // Parse the expression.
-    ExprResult Expr = ParseExpression(); // expression
+    ExprResult Expr; // expression
+    {
+      EnterExpressionEvaluationContext Consteval(
+          Actions, Sema::ExpressionEvaluationContext::ConstantEvaluated,
+          /*LambdaContextDecl=*/nullptr,
+          Sema::ExpressionEvaluationContextRecord::EK_Other,
+          /*ShouldEnter=*/CK == Sema::ConditionKind::ConstexprIf);
+
+      // Parse the expression.
+      Expr = ParseExpression(); // expression
+    }
     if (Expr.isInvalid())
       return Sema::ConditionError();
 
@@ -2186,6 +2195,19 @@ Parser::ParseCXXCondition(StmtResult *InitStmt, SourceLocation Loc,
   bool CopyInitialization = isTokenEqualOrEqualTypo();
   if (CopyInitialization)
     ConsumeToken();
+
+  Sema::ExpressionEvaluationContext NewEEC =
+      Actions.ExprEvalContexts.back().Context;
+  bool RuntimeEvaluated = Actions.ExprEvalContexts.back().IsRuntimeEvaluated;
+  if (DS.getTypeQualifiers() == DeclSpec::TQ_const)
+    RuntimeEvaluated = false;
+
+  if (CK == Sema::ConditionKind::ConstexprIf || DS.hasConstexprSpecifier()) {
+    RuntimeEvaluated = false;
+    NewEEC = Sema::ExpressionEvaluationContext::ConstantEvaluated;
+  }
+  EnterExpressionEvaluationContext Initializer(Actions, NewEEC, DeclOut);
+  Actions.ExprEvalContexts.back().IsRuntimeEvaluated = RuntimeEvaluated;
 
   ExprResult InitExpr = ExprError();
   if (getLangOpts().CPlusPlus11 && Tok.is(tok::l_brace)) {